# docs/swagger.yaml
basePath: /
definitions:
  v1.CreateUserRequest:
    properties:
      captcha_id:
        type: string
      captcha_solution:
        type: string
      gender:
        type: integer
      mail:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
  v1.CreateUserResponse:
    properties:
      avatar:
        type: string
      gender:
        type: integer
      mail:
        type: string
      name:
        type: string
      phone:
        type: string
      user_id:
        type: integer
    type: object
  v1.GetUserInfoResponse:
    properties:
      avatar:
        type: string
      gender:
        type: integer
      mail:
        type: string
      name:
        type: string
      phone:
        type: string
      user_id:
        type: integer
    type: object
host: api.shunjiecloud.com
info:
  contact:
    email: zhou_shengjie@outlook.com
    name: zhoushengjie
    url: https://www.shunjiecloud.com
  description: 瞬捷云 account-api
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://www.shunjiecloud.com
  title: 瞬捷云账号服务
  version: "1.0"
paths:
  /account/v1/users:
    get:
      description: 获取用户信息
      parameters:
      - description: 用户id，不带userId，则获取自己的userInfo
        in: query
        name: userId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.GetUserInfoResponse'
      summary: 获取用户信息
      tags:
      - users
    post:
      consumes:
      - application/json
      description: 创建用户
      parameters:
      - description: 密码先sha1，之后使用公钥加密，再stdBase64编码。
        in: body
        name: '{object}'
        required: true
        schema:
          $ref: '#/definitions/v1.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.CreateUserResponse'
      summary: 创建用户
      tags:
      - users
swagger: "2.0"